Extract the getFavicon function
