+++ packages/api/src/helpers/getFavicon.ts
@@ -1,0 +1,11 @@
+export const getFavicon = () => {
+  let favicon = document.querySelector("link[rel*='icon']")?.getAttribute('href');
+  if (favicon) {
+    try {
+      new URL(favicon);
+    } catch (e) {
+      favicon = window.location.origin + favicon;
+    }
+  }
+  return favicon;
+};
